pgsql: Move keyvalue to keyvalue module
